fix(Car) : InputValidator의 로직을 Car 클래스로 이동
